Finished prepare_calculation()

+++ include/ceval.h
@@ -1,6 +1,6 @@
 #ifndef _CEVAL_H
 #define _CEVAL_H
 
-long double ceval(char *text);
+double ceval(char *text);
 
 #endif
